oct25/langgraph/mailtoner/pyproject.toml
[project]
name = "mailtoner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.0.1",
    "langchain>=1.0.2",
    "langchain-google-vertexai>=3.0.1",
    "langgraph>=1.0.1",
    "python-dotenv>=1.1.1",
]
